Fixed Seed Packet

+++ src/main/java/net/cwserver/netty/packet/server/Packet16Join.java
@@ -6,17 +6,17 @@ import net.cwserver.netty.packet.CubeWorldPacket;
 @CubeWorldPacket.Packet(id = 16)
 public class Packet16Join extends CubeWorldPacket {
 
-    int entId;
+    long entId;
     byte[] connInfo;
 
-    public Packet16Join(int entId) {
+    public Packet16Join(long entId) {
         this.entId = entId;
         this.connInfo = new byte[4456];
     }
 
     @Override
     public void encode(ByteBuf buf) {
-        buf.writeInt(entId);
+        buf.writeLong(entId);
         buf.writeBytes(connInfo);
     }
 }
